Configuring IPv6 link-local address must remove the EUI64 link-local
The auto-generated EUI64 link-local address is not removed from the
interface after an IPv6 link-local address has been configured if there
is a large delay between admin up and link up. The reason is that the
check for existing link-locals has completed before the EUI64 LL is
added to the interface due to link up. The solution is not to generate
an EUI64 LL by setting the addr_gen_mode kernel parameter.

VRVDR-54643