Removing IPv6 LL address triggers a commit error
The commit 56fb10fc9c57 ("Configuring IPv6 link-local address must
remove the EUI64 link-local") introduced the use of addr_gen_mode.
When reverting to EUI64 mode, the LL is now autogenerated and so adding
this explicitly is no longer required, and depending on timing will
result in an error that it already exists. The solution is to remove
the explicit handling.

VRVDR-54727